Date Time Parsing Improvements (#119)
* Change date time parsing to allow dates and date time with days and months that are not two digits wide (i.e. allow 2020-1-1 instead of just 2020-01-01)
Change date time parsing to allow '/' in place of '-' in dates (note that outside of range queries the date string must be quoted or the / must be escaped)  (i.e. date:"2020/01/01" or date:2020\/01/01)
Allowing partial dates to be treated as a range i.e. date:2020-01 translates to date:[2020-01-01T00:00:00.000Z TO 2020-01-31T11:59:59.999Z].  Note that previously date:2020-01-01 would have translated to exactly 2020-01-01T00:00:00.000Z, and now it is translated to any time in the that day.
Add some test cases to test parsing and querying

* update so exclusive works with partial dates better (i.e. date:[2012 TO 2014} will now exclude any dates from 2014 and date:{2020-02-01 TO 2015] would exclude any time during the day 2020-02-01.)