bump mongo driver to 5.5.1 from 5.4.0 (performance gains in theory, but probably not much for the zulia use case) bump common compress to 1.28.0 from 1.27.1, fixes CVE-2025-48924
upgrade micronaut to 4.9.1 from 4.7.6 (virtual threads can be enabled, although experimental now https://micronaut.io/2025/06/30/transitioning-to-virtual-threads-using-the-micronaut-loom-carrier/)
upgrade protobuf to 4.31.1 from 3.25.5
upgrade grpc to 1.73.0 from 1.72.0
import DDSketch from the ddsketch repo and disable using the version from the jar in the build with the protobuf syntax in the gradle to get around the poisoning due to the CVE until they fix it.
Explicitly place the zulia commons project at the top of the classpath to use our version of ddsketches first.